+++ cloudnet/src/main/java/de/dytanic/cloudnet/database/IDatabase.java
@@ -40,7 +40,7 @@ public interface IDatabase extends INameable, AutoCloseable {
 
     void clear();
 
-    int getDocumentsCount();
+    long getDocumentsCount();
 
 
     ITask<Boolean> insertAsync(String key, JsonDocument document);
@@ -67,7 +67,7 @@ public interface IDatabase extends INameable, AutoCloseable {
 
     ITask<Void> clearAsync();
 
-    ITask<Integer> getDocumentsCountAsync();
+    ITask<Long> getDocumentsCountAsync();
 
 
 }
